Avoid downloading dev export if there's alread an up to date file

+++ scripts/get_db_export.sh
@@ -8,19 +8,35 @@
 export_file_zip="wca-developer-database-dump.zip"
 export_file_sql="wca-developer-database-dump.sql"
 
+download=true
+
 # First we check if the file exists.
 if [ -f $export_file_zip ]; then
-    echo "There's already a zip in this folder. Deleting."
-    rm $export_file_zip;
+    one_week=$(date -d 'now - 7 days' +%s) # export older than 1 week is suggested to be replaced
+    date_of_export=$(date -r "$export_file_zip" +%s)
+
+    if (( date_of_export <= one_week )); then # Here we check how old the export is.
+        echo "$export_file_zip is older than 7 days."
+        echo "Cleaning old data."
+        
+        rm $export_file_zip;
+        download=true;
+        
+    else
+        download=false
+        echo "$export_file_zip is up to date."
+    fi
+fi
+
+if [ "$download" = true ]; then
+    echo "Downloading the latest export."
+    wget -q https://www.worldcubeassociation.org/wst/wca-developer-database-dump.zip
 fi
 
 if [ -f $export_file_sql ]; then
     echo "There's already a sql in this folder. Deleting."
     rm $export_file_sql;
 fi
-
-echo "Downloading the latest export."
-wget -q https://www.worldcubeassociation.org/wst/wca-developer-database-dump.zip
 
 echo "Extracting $export_file_zip"
 unzip "$export_file_zip"
